fix: moving csv to csv_parser because of weird errors
