+++ app/tests/csv_parser/tests_csv_column.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from app.csv import CsvColumnStats
+from app.csv_parser import CsvColumnStats
 
 
 class CsvColumnTests(TestCase):
