+++ app/tests/csv_parser/tests_csv_utils.py
@@ -1,7 +1,7 @@
 import io
 from typing import Tuple
 from unittest import TestCase
-from app.csv import CsvColumnStats, derive_csv_column_types
+from app.csv_parser import CsvColumnStats, derive_csv_column_types
 
 
 class CsvUtilsTest(TestCase):
